Legal

+++ app/src/main/java/com/example/liftrix/data/local/dao/ContentReportsDao.kt
@@ -199,7 +199,7 @@ interface ContentReportsDao {
             val entity = contentReport.copy(
                 isDirty = false,
                 isSynced = true,
-                syncVersion = System.currentTimeMillis().toInt()
+                syncVersion = System.currentTimeMillis()
             )
             _insert(entity)
         }
