+++ app/src/main/java/com/example/liftrix/data/local/dao/GymBuddyDao.kt
@@ -212,7 +212,7 @@ interface GymBuddyDao {
             val entity = gymBuddy.copy(
                 isDirty = false,
                 isSynced = true,
-                syncVersion = System.currentTimeMillis().toInt()
+                syncVersion = System.currentTimeMillis()
             )
             _insert(entity)
         }
