+++ app/src/main/java/com/example/liftrix/data/local/dao/SettingsDao.kt
@@ -245,7 +245,7 @@ interface SettingsDao {
             val entity = settings.copy(
                 isDirty = false,
                 isSynced = true,
-                syncVersion = System.currentTimeMillis().toInt()
+                syncVersion = System.currentTimeMillis()
             )
             _insert(entity)
         }
