+++ app/src/main/java/com/example/liftrix/data/local/dao/SupportTicketDao.kt
@@ -253,7 +253,7 @@ interface SupportTicketDao {
             val entity = supportTicket.copy(
                 isDirty = false,
                 isSynced = true,
-                syncVersion = System.currentTimeMillis().toInt()
+                syncVersion = System.currentTimeMillis()
             )
             _insert(entity)
         }
