+++ app/src/main/java/com/example/liftrix/data/local/entity/ChatPreferencesEntity.kt
@@ -63,7 +63,7 @@ data class ChatPreferencesEntity(
     val isSynced: Boolean = false,
     
     @ColumnInfo(name = "sync_version", defaultValue = "0")
-    val syncVersion: Int = 0,
+    val syncVersion: Long = 0L,
     
     @ColumnInfo(name = "is_dirty", defaultValue = "0")
     val isDirty: Boolean = false,
