+++ app/src/main/java/com/example/liftrix/data/local/entity/FollowRequestEntity.kt
@@ -92,7 +92,7 @@ data class FollowRequestEntity(
     val isSynced: Boolean = false,
 
     @ColumnInfo(name = "sync_version", defaultValue = "0")
-    val syncVersion: Int = 0,
+    val syncVersion: Long = 0L,
 
     @ColumnInfo(name = "is_dirty", defaultValue = "0")
     val isDirty: Boolean = false,
