+++ app/src/main/java/com/example/liftrix/data/local/entity/WorkoutTemplateEntity.kt
@@ -70,7 +70,7 @@ data class WorkoutTemplateEntity(
     val isSynced: Boolean = false,
     
     @ColumnInfo(name = "sync_version", defaultValue = "1")
-    val syncVersion: Int = 1,
+    val syncVersion: Long = 1L,
     
     @ColumnInfo(name = "is_dirty", defaultValue = "0")
     val isDirty: Boolean = false,
